Renamed dir to 'ghc-filesystem', updated clang-format-hook-ignore, added origin record
