+++ .clang-format-hook-ignore
@@ -4,3 +4,5 @@
 cpp/src/edn-cpp/edn.hpp
 cpp/src/edn-cpp/example.cpp
 cpp/src/edn-cpp/repl.cpp
+
+cpp/src/ghc-filesystem/filesystem.hpp
